requirements.txt
manifest-ml[all]==0.1.8
pandas>=2.0.0
sqlalchemy<2.0.0
transformers
datasets==2.11.0
jsonlines>=3.1.0
sqlglot==11.5.5
click
rich
nltk>=3.5,<3.6
sqlparse
pebble
sentencepiece
duckdb==0.9.2
structlog==22.3.0
tensorboard==2.14.0
lightning==2.1.0
lmdb==1.4.1
cloudpathlib==0.13.0
peft==0.6.0
packaging==23.2
ninja==1.11.1.1
flash-attn
langchain
pydantic>2
packaging
